// angular-src/src/models/Spec.ts
/*
 * @Description: Model for Spec
 * @Author: Guozhi Tang
 * @Date: 2019-07-24 10:50:13
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:45:01
 */
export class Spec {
  className: String;
  moduleName: String;
  description: String;
  name: String;
  id: String;
}